[chore] Clean up constants file
Lots of random constants that didn't seem to ever get used or were
needed were left in here. Clean them out.
