+++ src/utils/constants.ts
@@ -7,56 +7,17 @@ export const networkName = 'cash-livenet'
 export const displayNetwork = 'livenet'
 
 // Wallet constants
-export const numAddresses = 10
-export const numChangeAddresses = 10
 export const recomendedBalance = 500_000
 export const nUtxoGoal = 10
-export const feeUpdateTimerMilliseconds = 60_000
 export const defaultFeePerByte = 2
 
 // Registry constants
 export const registrys = ['https://mainnet-keyserver.cashweb.io']
 
 // Relay constants
-export const pingTimeout = 20_000
-export const relayReconnectInterval = 10_000
 export const defaultAcceptancePrice = 100
 export const defaultRelayUrl = 'https://mainnet-relay.cashweb.io'
 export const relayUrlOptions = ['https://mainnet-relay.cashweb.io']
-export const defaultRelayData: {
-  profile: {
-    name: string
-    bio: string
-    avatar: string
-    pubKey?: Uint8Array
-  }
-  inbox: {
-    acceptancePrice?: number
-  }
-  notify: boolean
-} = {
-  profile: {
-    name: '',
-    bio: '',
-    avatar: '',
-  },
-  inbox: {
-    acceptancePrice: defaultAcceptancePrice,
-  },
-  notify: true,
-}
-export const pendingRelayData = {
-  profile: {
-    name: 'Loading...',
-    bio: '',
-    avatar: '',
-  },
-  inbox: {
-    acceptancePrice: NaN,
-  },
-  notify: true,
-  lastUpdateTime: 0,
-}
 
 // Avatar constants
 export const defaultAvatars = [
@@ -81,9 +42,6 @@ export const defaultContacts = [
     address: 'lotus_16PSJPAVocAM5behRWxqwQnpEVRPJrV4XxbthBhJR',
   },
 ]
-
-// Notification constants
-export const notificationTimeout = 4000
 
 // Contact defaults
 export const defaultUpdateInterval = 1000 * 60 * 60 * 1 * 1
